test repositories

+++ backend/src/main/java/com/example/germanteaching/auth/entity/PasswordResetToken.java
@@ -10,7 +10,7 @@ public class PasswordResetToken {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column (name = "id")
-    private Long id;
+    private Integer id;
 
     @Column (name = "token", nullable =false, length = 255)
     private String token;
@@ -64,10 +64,10 @@ public class PasswordResetToken {
 
 
     // Getters and Setters
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
-    public void setId(long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
